Move some files to new utils-dir
Make the root src directory more easy to read and maybe understand.
