+++ src/utils/ai_predict.ts
