+++ src/utils/dbUtils.ts
@@ -1,7 +1,7 @@
 import { Pool, PoolClient } from 'pg';
 
-import { ApiError } from './utils';  // TODO: Don't use ./utils.ts because of ./index-debug.ts
-import { SpraxAPIdbCfg, UserAgent, Skin, MinecraftUser, Cape, CapeType, MinecraftProfile } from './global';
+import { ApiError } from '../utils';  // TODO: Don't use ./utils.ts because of ./index-debug.ts
+import { SpraxAPIdbCfg, UserAgent, Skin, MinecraftUser, Cape, CapeType, MinecraftProfile } from '../global';
 
 export class dbUtils {
   private pool: Pool | null = null;
